Included parameters for Scope on newly created parameters bank global variable (g_param_bank)
Created new type of parameter: pointer to float.

Fixed bugs on timeslicer initializations.

+++ app/communication_drivers/common/timeslicer.c
@@ -20,6 +20,7 @@
  *
  */
 
+#include <math.h>
 #include "communication_drivers/common/timeslicer.h"
 
 void init_timeslicer(timeslicer_t *p_ts, float freq_base)
@@ -33,7 +34,15 @@ void init_timeslicer(timeslicer_t *p_ts, float freq_base)
 void cfg_timeslicer(timeslicer_t *p_ts, float freq_sampling)
 {
     p_ts->freq_ratio = (uint16_t) roundf(p_ts->freq_base.f / freq_sampling);
-    p_ts->freq_sampling.f = p_ts->freq_base.f / p_ts->freq_ratio;
+
+    /**
+     *  TODO: Investigate one-bit accuracy-error in some cases, like:
+     *          freq_base = 48000.0
+     *          freq_ratio = 480
+     *          freq_sampling = 100.000008
+     */
+    p_ts->freq_sampling.f = p_ts->freq_base.f / ((float) p_ts->freq_ratio);
+
     p_ts->counter = p_ts->freq_ratio;
 }
